% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesize.R
\name{gesize}
\alias{gesize}
\title{Genome size studies}
\usage{
gesize(n, m)
}
\arguments{
\item{symbol}{}
}
\description{
Calculates the GC percentage of given sequences and calculates the length of them (aka number of bases).
}
