% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meancat.R
\name{meancat}
\alias{meancat}
\title{Mean of a value according to a category}
\usage{
meancat(n, m)
}
\description{
Calculates the mean of a numerical variable (n) according to certain category (m) --> meancat(n, m)
}
