% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percencat.R
\name{percencat}
\alias{percencat}
\title{Percentages of categories}
\usage{
percencat(n)
}
\arguments{
\item{symbol}{}
}
\description{
Calculates the percentage of representation of a given category present on a string.
}
