% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_data}
\alias{random_data}
\title{random_data}
\format{
A data frame with 9 rows and 3 variables:
\describe{
  \item{person}{Hypothetical person name, person1, person2, person3, etc..}
  \item{country}{Hypothetical country.}
  \item{n_dogs}{Hypothetical number of dogs.}
}
}
\source{
<https://www.github.com/alejandrorgijon/plada>
}
\usage{
random_data
}
\description{
Randomly generated dataset
}
\keyword{datasets}
