% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sequences}
\alias{sequences}
\title{sequences}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{sequences}{Sequences names, seq1, seq2, seq3, seq4 and seq5.}
  \item{genseqs}{Hypothetical genetic sequences.}
}
}
\source{
<https://www.github.com/alejandrorgijon/plada>
}
\usage{
sequences
}
\description{
Randomly generated genetic sequences.
}
\keyword{datasets}
