% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxabu.R
\name{taxabu}
\alias{taxabu}
\title{Taxonomic categorization effort analysis}
\usage{
taxabu(df)
}
\arguments{
\item{symbol}{}
}
\description{
Calculates how many mOTUs are classified until different taxonomical levels.
The dataframe should contain 8 columns, normally in the order as follows: mOTU, domain, phyla, class, order,
family, genera, species.
}
