% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_envi.R
\name{which_envi}
\alias{which_envi}
\title{Representation of a variable in another variable}
\usage{
which_envi(category1, category2)
}
\arguments{
\item{symbol}{}
}
\description{
Calculates if a given categoric variable has a representation of above or under 50% according to another categoric variable. n and m represents the 2 categoric variables you want to analyse. o represents the dataframe you want to analyse.
}
